cleanup to interface
moved open_connection and close_connection to  interface.c file, added
write_header and write_data helper functions, added some debugging to
gtec code (for which I want the new helper functions), some small
cleanups
