+++ realtime/src/buffer/src/message.h
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright (C) 2008, Robert Oostenveld & Christian Hesse
  * F.C. Donders Centre for Cognitive Neuroimaging, Radboud University Nijmegen,
  * Kapittelweg 29, 6525 EN Nijmegen, The Netherlands
@@ -56,7 +56,7 @@ typedef uint64_t UINT64_T;
 
 /* these can be used for indexing the buffer pointer as array */
 #define WORDSIZE_CHAR    sizeof(CHAR_T   )
-#define WORDSIZE_UINT8   sizeof(UINT8_T  ) 
+#define WORDSIZE_UINT8   sizeof(UINT8_T  )
 #define WORDSIZE_UINT16  sizeof(UINT16_T )
 #define WORDSIZE_UINT32  sizeof(UINT32_T )
 #define WORDSIZE_UINT64  sizeof(UINT64_T )
@@ -120,7 +120,7 @@ typedef uint64_t UINT64_T;
 #define EVENTSEL_MINSAMPLE 4
 #define EVENTSEL_MAXSAMPLE 5
 
-/*  
+/*
   if event->def->sample == EVENT_AUTO_SAMPLE, automatically insert current sample index instead
 */
 #define EVENT_AUTO_SAMPLE   -1
@@ -142,7 +142,7 @@ enum {
     /** FT_CHUNK_CHANNEL_FLAGS contains a 0-terminated string describing the type of flags,
         and after that N (=#channels) bytes describing each channel. This is useful for
         specifying that a channel can have a discrete number of different types, e.g.
-        chunk_data = "meg_ad_eog\0\1\1\1\1\3\3\2\2" should be used for a system with 8 channels, 
+        chunk_data = "meg_ad_eog\0\1\1\1\1\3\3\2\2" should be used for a system with 8 channels,
         the first four of which are for MEG, then 2 channels EOG, then 2 channels A/D.  */
     FT_CHUNK_CHANNEL_FLAGS = 2,
 
@@ -160,16 +160,16 @@ enum {
 
     /** FT_CHUNK_SIEMENS_AP contains Siemens Protocol data in ASCII format (string) */
     FT_CHUNK_SIEMENS_AP = 6,
-    
+
     /** FT_CHUNK_CTF_RES4 contains a .res4 file as written by the CTF MEG acquisition software (binary) */
     FT_CHUNK_CTF_RES4 = 7,
-    
+
     /** FT_CHUNK_NEUROMAG_HEADER contains a .fif file as written by the Neuromag MEG acquisition software (binary) */
     FT_CHUNK_NEUROMAG_HEADER = 8,
-    
+
     /** FT_CHUNK_NEUROMAG_ISOTRAK contains a .fif file as written by the Neuromag MEG acquisition software (binary) */
     FT_CHUNK_NEUROMAG_ISOTRAK = 9,
-    
+
     /** FT_CHUNK_NEUROMAG_HPIRESULT contains a .fif file as written by the Neuromag MEG acquisition software (binary) */
     FT_CHUNK_NEUROMAG_HPIRESULT = 10
 };
@@ -206,8 +206,8 @@ typedef struct {
     UINT32_T type_type;   /* usual would be DATATYPE_CHAR */
     UINT32_T type_numel;  /* length of the type string */
     UINT32_T value_type;
-    UINT32_T value_numel; 
-    INT32_T sample; 
+    UINT32_T value_numel;
+    INT32_T sample;
     INT32_T offset;
     INT32_T duration;
     UINT32_T bufsize;     /* size of the buffer in bytes */
