+++ realtime/src/buffer/src/util.c
@@ -20,7 +20,7 @@
 unsigned int bufread(int s, void *buf, unsigned int numel) {
 		unsigned int numcall = 0, numread = 0, verbose = 0;
 		int numthis = 0;
-		
+
 		while (numread<numel) {
 
 				numthis = recv(s, (char*)buf+numread, numel-numread, 0);
@@ -35,10 +35,6 @@ unsigned int bufread(int s, void *buf, unsigned int numel) {
 						fprintf(stderr, "bufread: read %d bytes\n", numthis);
 				numread += numthis;
 				numcall ++;
-#ifndef PLATFORM_WIN32			/* SK: I think this shouldn't be necessary on any platform: the sockets are blocking */
-				if (numread<numel)
-						usleep(1000);
-#endif
 		}
 		if (verbose>1)
 				fprintf(stderr, "bufread: reading the complete buffer required %d calls\n", numcall);
@@ -66,10 +62,6 @@ unsigned int bufwrite(int s, const void *buf, unsigned int numel) {
 						fprintf(stderr, "bufwrite: wrote %d bytes\n", numthis);
 				numwrite += numthis;
 				numcall ++;
-#ifndef PLATFORM_WIN32			/* SK: I think this shouldn't be necessary on any platform: the sockets are blocking */				
-				if (numwrite<numel)
-						usleep(1000);
-#endif
 		}
 		if (verbose>1)
 				fprintf(stderr, "bufwrite: writing the complete buffer required %d calls\n", numcall);
@@ -110,108 +102,6 @@ unsigned int append(void **buf1, unsigned int bufsize1, void *buf2, unsigned int
 		return (bufsize1+bufsize2);
 }
 
-int close_connection(int s) {
-		int status = 0, verbose = 0;
-		if (verbose>0)
-				fprintf(stderr, "close_connection: socket = %d\n", s);
-		if (s>0)
-				status = closesocket(s);	/* it is a TCP connection */
-		if (status!=0)
-				perror("close_connection");
-		return status;
-}
-
-int open_connection(const char *hostname, int port) {
-		int verbose = 0;
-		int s, retry;
-		struct sockaddr_in sa;
-		struct hostent *host;
-#ifdef WIN32
-		static WSADATA wsa = {0,0}; /* check version fields to only initialise once */
-#endif
-
-		if (port==0) {
-				if (verbose>0)
-						fprintf(stderr, "open_connection: using direct memory copy\n");
-				return 0;
-		}
-		else {
-				if (verbose>0)
-						fprintf(stderr, "open_connection: server = %s, port = %d\n", hostname, port);
-		}
-
-#ifdef WIN32
-		if (wsa.wVersion == 0) {
-			/* We only need to do this once ... and actually have a corresponding WSACleanup call somewhere */
-			if(WSAStartup(MAKEWORD(1, 1), &wsa)) {
-				fprintf(stderr, "open_connection: cannot start sockets\n");
-				/* FIXME should this exception be handled more explicitely?  */
-			}
-		}
-#endif
-
-		if ((host = gethostbyname(hostname)) == NULL) {
-				fprintf(stderr, "open_connection: nslookup1 failed on '%s'\n", hostname);
-				return -1;
-		}
-
-		if (host->h_length == 0) {
-				fprintf(stderr, "open_connection: nslookup2 failed on '%s'\n", hostname);
-				return -1;
-		}
-
-		bzero(&sa, sizeof sa);
-		sa.sin_family = AF_INET;
-		sa.sin_port = htons(port);
-
-		memcpy(&(sa.sin_addr.s_addr), host->h_addr_list[0], sizeof(sa.sin_addr.s_addr));
-
-		if ((s = socket(PF_INET, SOCK_STREAM, 0)) < 0) {
-				if (verbose>0)
-						fprintf(stderr, "open_connection: socket = %d\n", s);
-				perror("open_connection");
-				return -1;
-		}
-    
-		retry = 10;
-		while (retry>0) {
-				if (connect(s, (struct sockaddr *)&sa, sizeof sa)<0) {
-						/* wait 5 miliseconds and try again */
-						usleep(5000);
-						retry--;
-				}
-				else {
-						/* this signals that the connection has been made */
-						retry = -1;
-				}
-		}
-		if (retry==0) {
-				/* close the socket */
-				closesocket(s);
-				/* it failed on mutliple attempts, give up */
-				return -2;
-		}
-
-		/*
-		   while (connect(s, (struct sockaddr *)&sa, sizeof sa) < 0) {
-		   perror("open_connection connect");
-		   usleep(1000000);
-		   }
-		 */
-
-		if (verbose>0)
-				fprintf(stderr, "open_connection: connected to %s:%d on socket %d\n", hostname, port, s);
-
-#ifdef DISABLE_NAGLE
-		{
-			int optval = 1;
-			setsockopt(s, IPPROTO_TCP, TCP_NODELAY, &optval, sizeof(optval));
-		}
-#endif
-        
-		return s;
-}
-
 void check_datatypes() {
 		/* check datatypes */
 		if (WORDSIZE_CHAR    !=1) { fprintf(stderr, "invalid size of CHAR    (%d)\n", (int) WORDSIZE_CHAR   ); exit(-1); }
@@ -233,19 +123,17 @@ void check_datatypes() {
 		if (sizeof(eventsel_t)    !=8 ) { fprintf(stderr, "invalid size of eventsel_t  \n"); exit(-1); }
 }
 
-
-
 unsigned int wordsize_from_type(UINT32_T data_type) {
 	switch(data_type) {
 		case DATATYPE_CHAR:
 			return WORDSIZE_CHAR;
-		case DATATYPE_UINT8:		
+		case DATATYPE_UINT8:
 		case DATATYPE_INT8:
 			return WORDSIZE_INT8;
 		case DATATYPE_UINT16:
 		case DATATYPE_INT16:
 			return WORDSIZE_INT16;
-		case DATATYPE_UINT32:				
+		case DATATYPE_UINT32:
 		case DATATYPE_INT32:
 			return WORDSIZE_INT32;
 		case DATATYPE_UINT64:
@@ -264,7 +152,7 @@ const ft_chunk_t *find_chunk(const void *buf, unsigned int offset0, unsigned int
 	while (bufpos + sizeof(ft_chunkdef_t) <= size) {
 		const ft_chunk_t *chunk = (ft_chunk_t *) ((char *)buf + bufpos);
 		if (chunk->def.type == chunk_type) return chunk;
-		
+
 		bufpos += sizeof(ft_chunkdef_t) + chunk->def.size;
 	}
 	return NULL;
@@ -283,26 +171,26 @@ const ft_chunk_t *find_chunk(const void *buf, unsigned int offset0, unsigned int
 int check_event_array(unsigned int size, const void *buf) {
 	unsigned int offset=0;
 	int numEvents = 0;
-	
+
 	while (offset + sizeof(eventdef_t) <= size) {
-		const eventdef_t *E;	
+		const eventdef_t *E;
 		unsigned int wsType, wsValue;
-		
+
 		/* Set our event pointer to the current location within the array */
 		E = (const eventdef_t *) ((char *) buf + offset);
-		
+
 		/* Increase the offset by the size of this event, and check whether it's fully
 		   contained within the given array.
 		*/
 		offset += sizeof(eventdef_t) + E->bufsize;
 		if (offset > size) goto error;
-		
+
 		/* Check whether "type" and "value" are of known type */
 		wsType = wordsize_from_type(E->type_type);
 		if (wsType == 0) goto error;
 		wsValue = wordsize_from_type(E->value_type);
 		if (wsValue == 0) goto error;
-		
+
 		/* Check whether "type" and "value" are contained in this event's "buf" */
 		if (wsType * E->type_numel + wsValue * E->value_numel > E->bufsize) goto error;
 
@@ -313,49 +201,3 @@ int check_event_array(unsigned int size, const void *buf) {
 error:
 	return -(1+numEvents);
 }
-
-
-#ifdef WIN32
-int open_unix_connection(const char *name) {
-	return -1;
-}
-#else
-int open_unix_connection(const char *name) {
-	int verbose = 0;
-	int s, retry;
-	struct sockaddr_un sa;
-
-	bzero(&sa, sizeof(sa));
-	sa.sun_family = AF_UNIX;
-	strncpy(sa.sun_path, name, sizeof(sa.sun_path));
-		
-	s = socket(AF_UNIX, SOCK_STREAM, 0);
-	if (s < 0) {
-		perror("open_unix_connection, socket");
-		return -1;
-	}
-
-	retry = 10;
-	while (retry>0) {
-		if (connect(s, (struct sockaddr *)&sa, sizeof(sa))<0) {
-			/* wait 5 miliseconds and try again */
-			perror("open_connection");
-			usleep(5000);
-			retry--;
-		} else {
-			/* this signals that the connection has been made */
-			retry = -1;
-			}
-		}
-	if (retry==0) {
-		/* it failed on mutliple attempts, give up */
-		return -2;
-	}
-
-	if (verbose>0)
-		fprintf(stderr, "open_unix_connection: connected to %s on socket %d\n", name,  s);
-        
-	return s;
-}
-#endif
-
